Refactor cardinal direction and make it listen to the currently shown texture
